+++ test/dummy/app/controllers/api/v1/everything_controller.rb
@@ -47,7 +47,7 @@ class API::V1::EverythingController < API::V1::BaseController
 
     param :date_time_iso8601, :date_time_iso8601
 
-    param :date_time_posix, :date_time_posix
+    param :date_time_unix_epoch, :date_time_unix_epoch
 
     param :enum1, [:enum, %w[one two three]]